Move parsed sample sheet DTOs from src into test
Ticket:https://github.com/SiLeBAT/mibi-portal-server/issues/463
